Purge build deps to shrink image from 2.32 GB to 1.76 GB
The space savings are for the 12.15 image, built locally before and
after this commit.

Combine the pg and pg_cron layers to avoid installing & purging
build-depends several times.

Gosu is used in entrypoint, so it is not purged.
locales is required by Pg at runtime, so it is not purged.