Allow pg_cron to work on any configured POSTGRES_DB (#95)
Due to incorrect ownership of /etc/postgresql/custom-postgresql.conf
the pg_cron.sh script was not able to add POSTGRES_DB to the
`cron.database_name` config. Thus the only DB that would work with
pg_cron is the 'circle_test' DB, as this was hard-coded in the example
file.

This fixes ownership of the files in the /etc/postgresql directory, so
the pg_cron.sh script can run as expected.

Also remove the cron-related lines from custom-postgresql.conf since
pg_cron.sh is adding those two lines and we don't want confusion
around duplicate settings. (Last one wins, for what it's worth.)